Backend/ml-services/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow==2.15.0
Pillow==10.0.1
numpy==1.24.3
joblib==1.3.2
matplotlib==3.7.2
kagglehub==0.2.0
